PSA eBPF: Core Test: Decrement
Decrements an incoming value by various amounts. The decremented
values are sent back to the sender. It indirectly tests the
StrictAssignVar command.

Note: A warning is produced during the eBPF C-Code compilation
because for a uint64 negative value is transformed in the p4c-ebpf
compilation step to an addition of a 2s complement number which is
then seen in clang as a very large signed integer type, which is
to large to be represented. The consequence is that this value is
interpreted as unsigned.