Simple Test for SetStandardField
Sets hdr.udp.len to 9, overwriting the correct length (10). The effect of this is, that only one of the two specified input chars is read.

This test is not a complete, as it only tests the effect of changing the UdpLen field. Currently, that is the only idea for testing this command from inside our existing test infrastructure though. Additional tests probably need to be implemented one level higher.